Add "superdrop" capability to spamfilter
This new action works the same as drop, however it applies even if the
target user or channel is +u. In other words, it is not possible to "opt
out" of the superdrop action. Opers remain immune, both as senders and
recipients, so that spamfilter cannot be abused to lock people out of
managing filters, talking to services, etc.